refactor(login): remove Auth component move Auth code to Login component
