--- src/modules/auth/components/auth.module.css
@@ -1,99 +1,0 @@
-.authContainer {
-	display: flex;
-	/* gap: 10px; */
-	align-items: center;
-	flex-direction: column;
-	justify-content: space-between;
-	width: 30%;
-	min-height: 400px;
-	margin: auto;
-	background: #fff;
-	box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-	border-radius: 5px;
-	margin-top: 350px;
-	padding: 35px 35px;
-	box-sizing: border-box;
-}
-
-.loginHeader {
-	font-size: 32px;
-	font-weight: 500;
-	line-height: 12px;
-	text-align: center;
-}
-
-.loginForm {
-	width: 100%;
-}
-
-.formItem {
-	display: flex;
-	flex-direction: column;
-	align-items: start;
-	width: 100%;
-}
-
-.nameOfItem {
-	display: block;
-	margin-top: 30px;
-	font-size: 16px;
-	font-weight: 500;
-}
-
-.sendLogin {
-	margin-top: 50px;
-	margin-bottom: 10px;
-}
-
-@media (min-width: 320px) {
-	.authContainer {
-		width: 100%;
-		margin-top: 0px;
-		height: 85vh;
-	}
-
-	.loginHeader {
-		margin-top: 50px;
-		margin-bottom: 50px;
-	}
-}
-
-@media (min-width: 769px) {
-	.loginHeader {
-		margin-top: 100px;
-		margin-bottom: 120px;
-	}
-}
-
-@media (min-width: 1201px) {
-	.authContainer {
-		width: 30%;
-		margin-top: 300px;
-		height: 500px;
-	}
-
-	.loginHeader {
-		margin-top: 30px;
-		margin-bottom: 30px;
-	}
-}
-
-@media (min-width: 1500px) {
-	.authContainer {
-		margin-top: 300px;
-		width: 25%;
-	}
-}
-
-@media (min-width: 1800px) {
-	.authContainer {
-		margin-top: 300px;
-		width: 20%;
-	}
-}
-
-@media (min-width: 2100px) {
-	.authContainer {
-		width: 15%;
-	}
-}
